requirements/test.txt
behave
pytest==3.8.2
pytest-runner==4.2
pytest-mock
pytest-docker
pytest-flask
docker-compose
pyyaml
psycopg2-binary
sqlalchemy_utils